+++ src/main/java/io/github/mbenincasa/javaopenweathermapclient/client/DefaultOpenWeatherMapClient.java
@@ -7,6 +7,8 @@ import io.github.mbenincasa.javaopenweathermapclient.request.weatherForecast.Cli
 import io.github.mbenincasa.javaopenweathermapclient.request.weatherForecast.DailyWeatherForecastRequest;
 import io.github.mbenincasa.javaopenweathermapclient.request.weatherForecast.FiveDaysWeatherForecastRequest;
 import io.github.mbenincasa.javaopenweathermapclient.request.weatherForecast.HourlyWeatherForecastRequest;
+import io.github.mbenincasa.javaopenweathermapclient.request.weatherMaps.AdvancedWeatherMapRequest;
+import io.github.mbenincasa.javaopenweathermapclient.request.weatherMaps.BasicWeatherMapRequest;
 
 public class DefaultOpenWeatherMapClient implements OpenWeatherMapClient {
 
@@ -49,5 +51,15 @@ public class DefaultOpenWeatherMapClient implements OpenWeatherMapClient {
     @Override
     public GeocodingRequest geocoding() {
         return new GeocodingRequest(this.apiKey);
+    }
+
+    @Override
+    public BasicWeatherMapRequest basicWeatherMap() {
+        return new BasicWeatherMapRequest(this.apiKey);
+    }
+
+    @Override
+    public AdvancedWeatherMapRequest advancedWeatherMap() {
+        return new AdvancedWeatherMapRequest(this.apiKey);
     }
 }