+++ src/main/java/io/github/mbenincasa/javaopenweathermapclient/client/OpenWeatherMapClient.java
@@ -7,6 +7,8 @@ import io.github.mbenincasa.javaopenweathermapclient.request.weatherForecast.Cli
 import io.github.mbenincasa.javaopenweathermapclient.request.weatherForecast.DailyWeatherForecastRequest;
 import io.github.mbenincasa.javaopenweathermapclient.request.weatherForecast.FiveDaysWeatherForecastRequest;
 import io.github.mbenincasa.javaopenweathermapclient.request.weatherForecast.HourlyWeatherForecastRequest;
+import io.github.mbenincasa.javaopenweathermapclient.request.weatherMaps.AdvancedWeatherMapRequest;
+import io.github.mbenincasa.javaopenweathermapclient.request.weatherMaps.BasicWeatherMapRequest;
 
 public interface OpenWeatherMapClient {
 
@@ -23,4 +25,8 @@ public interface OpenWeatherMapClient {
     AirPollutionRequest airPollution();
 
     GeocodingRequest geocoding();
+
+    BasicWeatherMapRequest basicWeatherMap();
+
+    AdvancedWeatherMapRequest advancedWeatherMap();
 }